requirements.txt
# Copyright (C) 2023 Mitsubishi Electric Research Laboratories (MERL)
#
# SPDX-License-Identifier: AGPL-3.0-or-later

geoopt==0.5.0
librosa==0.10.0
matplotlib==3.7.1
PyQt6==6.4.2
pyqtgraph==0.13.2
pytorch_lightning==2.0.0
PyYAML==6.0
scipy==1.10.1
seaborn==0.12.2
soundfile==0.12.1
torch==2.0.0
torchaudio==2.0.1
